+++ mockpipe/__init__.py
@@ -3,4 +3,4 @@ from ._version import __version__
 from mockpipe.mockpipe import MockPipe
 from mockpipe.imposter import fake
 
-__all__ = ["MockPipe", "fake"]
+__all__ = ["MockPipe", "fake", "__version__"]